+++ aula15EXRC/Exer08.java
@@ -10,12 +10,13 @@ public class Exer08 {
 		Scanner scan = new Scanner(System.in);
 		
 		System.out.println("Me mande 3 preços DIFERENTES");
+		
 		System.out.println("Valor do primeiro produto: ");
-		int produto1 = scan.nextInt();
+		double produto1 = scan.nextDouble();
 		System.out.println("Valor do segundo produto: ");
-		int produto2 = scan.nextInt();
+		double produto2 = scan.nextDouble();
 		System.out.println("Valor do terceiro produto: ");
-		int produto3 = scan.nextInt();
+		double produto3 = scan.nextDouble();
 		
 		if (produto1 < produto2 && produto1 < produto3) {
 			System.out.println("O primeiro produto tem preço menor, você deve comprar.");
