+++ aula15EXRC/Exer11.java
@@ -29,8 +29,8 @@ public class Exer11 {
 		int salario = scan.nextInt();
 		
 		int percentual = 0;
-		int aumento = 0;
-		int novoSalario = 0;
+		double aumento = 0;
+		double novoSalario = 0;
 		
 		if (salario <= 280) {
 			percentual = 20;
@@ -49,9 +49,6 @@ public class Exer11 {
 		System.out.println("O reajuste é de " + percentual + "%");
 		System.out.println("Seu aumento é de " + aumento + "R$.");
 		System.out.println("Seu novo salário é de " + novoSalario + "R$.");
-		
-		
-		
 
 	}
